Move JWT Beans to separate config class
