+++ src/main/java/com/teamchallenge/bookti/config/SecurityConfig.java
@@ -1,22 +1,13 @@
 package com.teamchallenge.bookti.config;
 
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.JWKSet;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
-import com.nimbusds.jose.jwk.source.JWKSource;
-import com.nimbusds.jose.proc.SecurityContext;
 import com.teamchallenge.bookti.security.CustomUserDetailsService;
 import com.teamchallenge.bookti.security.handler.CustomAccessDeniedHandler;
 import com.teamchallenge.bookti.security.handler.CustomRestAuthenticationEntryPoint;
 import com.teamchallenge.bookti.security.jwt.JwtToAuthorizedUserConverter;
-import com.teamchallenge.bookti.security.jwt.KeyPairUtils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
@@ -28,11 +19,6 @@ import org.springframework.security.config.annotation.web.configurers.AbstractHt
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.JwtEncoder;
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
-import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationProvider;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
@@ -51,7 +37,6 @@ public class SecurityConfig {
     private final JwtToAuthorizedUserConverter jwtToUserConverter;
     private final CustomRestAuthenticationEntryPoint restAuthenticationEntryPoint;
     private final CustomAccessDeniedHandler accessDeniedHandler;
-    private final KeyPairUtils keyPairUtils;
     private final ApplicationProperties applicationProperties;
 
     @Bean
@@ -77,52 +62,6 @@ public class SecurityConfig {
                         .jwt(jwt -> jwt.jwtAuthenticationConverter(jwtToUserConverter))
                 )
                 .build();
-    }
-
-    @Bean
-    @Primary
-    public JwtDecoder jwtAccessTokenDecoder() {
-        return NimbusJwtDecoder
-                .withPublicKey(keyPairUtils.getAccessTokenPublicKey())
-                .build();
-    }
-
-    @Bean
-    @Primary
-    public JwtEncoder jwtAccessTokenEncoder() {
-        JWK jwk = new RSAKey
-                .Builder(keyPairUtils.getAccessTokenPublicKey())
-                .privateKey(keyPairUtils.getAccessTokenPrivateKey())
-                .build();
-        JWKSource<SecurityContext> jwkSource = new ImmutableJWKSet<>(new JWKSet(jwk));
-        return new NimbusJwtEncoder(jwkSource);
-    }
-
-    @Bean
-    @Qualifier("jwtRefreshTokenDecoder")
-    public JwtDecoder jwtRefreshTokenDecoder() {
-        return NimbusJwtDecoder
-                .withPublicKey(keyPairUtils.getRefreshTokenPublicKey())
-                .build();
-    }
-
-    @Bean
-    @Qualifier("jwtRefreshTokenEncoder")
-    public JwtEncoder jwtRefreshTokenEncoder() {
-        JWK jwk = new RSAKey
-                .Builder(keyPairUtils.getRefreshTokenPublicKey())
-                .privateKey(keyPairUtils.getRefreshTokenPrivateKey())
-                .build();
-        JWKSource<SecurityContext> jwkSource = new ImmutableJWKSet<>(new JWKSet(jwk));
-        return new NimbusJwtEncoder(jwkSource);
-    }
-
-    @Bean
-    @Qualifier("jwtRefreshTokenAuthenticationProvider")
-    public JwtAuthenticationProvider jwtRefreshTokenAuthenticationProvider() {
-        JwtAuthenticationProvider provider = new JwtAuthenticationProvider(jwtRefreshTokenDecoder());
-        provider.setJwtAuthenticationConverter(jwtToUserConverter);
-        return provider;
     }
 
     @Bean
